requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
executing==2.1.0
filelock==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.29.0
itemadapter==0.9.0
itemloaders==1.3.2
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
Protego==0.3.1
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyMuPDF==1.24.14
pyOpenSSL==24.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
queuelib==1.7.0
RapidFuzz==3.10.1
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.4.0
tldextract==5.1.2
tornado==6.4.1
traitlets==5.14.3
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
w3lib==2.2.1
wcwidth==0.2.13
zope.interface==7.1.1
